Moved defs around; removed SimpleWriter; renamed CanonicalSerialize to HashableLeaf
